+++ CHANGELOG.md
@@ -4,7 +4,8 @@ IMPROVEMENTS
 
 * Add `--heartbeat-interval` CLI flag and `MCP_HEARTBEAT_INTERVAL` env var for HTTP heartbeat in load-balanced environments
 * Set custom User-Agent header for TFE API requests to enable tracking MCP server usage separately from other go-tfe clients [268](https://github.com/hashicorp/terraform-mcp-server/pull/268)
-* Adding a new cli flags `--log-level` to set the desired log level for the server logs and `--log-format` for the logs formatting [286](https://github.com/hashicorp/terraform-mcp-server/pull/286)
+* Adding cli flags `--log-level` to set the desired log level for the server logs and `--log-format` for the logs formatting and corresponding env vars `LOG_LEVEL` and `LOG_FORMAT` [286](https://github.com/hashicorp/terraform-mcp-server/pull/286)
+* Adding cli flag `--disable-streaming` and env var `MCP_DISABLE_STREAMING` to disable Server-Sent Events (SSE) streaming and use direct HTTP response only [294]](https://github.com/hashicorp/terraform-mcp-server/pull/294)
 
 FIXES